Slice the request objects before before shifting them out of the original array.
The request object is required by code that is called after the code that shifts all items from the array.
This caused a bug in the sync code, where the lastSyncCount was not set, causing the system to think that there were changes on the remote while there were not.